+++ src/main/java/techcourse/myblog/service/LogInService.java
@@ -1,0 +1,7 @@
+package techcourse.myblog.service;
+
+import org.springframework.stereotype.Service;
+
+@Service
+public class LogInService {
+}
