+++ src/main/java/techcourse/myblog/web/LogInController.java
@@ -1,0 +1,19 @@
+package techcourse.myblog.web;
+
+import org.springframework.stereotype.Controller;
+import org.springframework.web.bind.annotation.GetMapping;
+import techcourse.myblog.service.LogInService;
+
+@Controller
+public class LogInController {
+    private LogInService logInService;
+
+    public LogInController(LogInService logInService) {
+        this.logInService = logInService;
+    }
+
+    @GetMapping("/login")
+    public String showLoginPage() {
+        return "login";
+    }
+}
