+++ src/main/resources/templates/fragments/logout-header.html
@@ -14,7 +14,7 @@
         </ul>
         <ul class="nav-right">
             <li class="user-profile">
-                <a href="/users/login">
+                <a href="/login">
                     <div class="user-info login-btn pointer mrg-right-10">
                         <span class="name pdd-right-5 text-white">Login</span>
                     </div>
