+++ src/main/resources/templates/login.html
@@ -48,7 +48,7 @@
                                                     </div>
                                                     <form>
                                                         <div class="form-group">
-                                                            <input type="email" class="form-control" placeholder="Id">
+                                                            <input type="email" class="form-control" placeholder="Email">
                                                         </div>
                                                         <div class="form-group">
                                                             <input type="password" class="form-control"
