+++ src/main/resources/templates/sign-up.html
@@ -69,7 +69,7 @@
                                                     </form>
                                                     <div class="text-dark">
                                                         <span>이미 회원이신가요?</span>
-                                                        <a href="/users/login"
+                                                        <a href="/login"
                                                            class="text-info pointer mrg-left-5">로그인</a>
                                                     </div>
                                                 </div>
