+++ src/test/java/techcourse/myblog/web/LogInControllerTest.java
@@ -1,0 +1,26 @@
+package techcourse.myblog.web;
+
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.web.server.LocalServerPort;
+import org.springframework.test.web.reactive.server.WebTestClient;
+
+@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
+class LogInControllerTest {
+    @LocalServerPort
+    int randomPortNumber;
+
+    @Autowired
+    WebTestClient webTestClient;
+
+    @Test
+    @DisplayName("로그인 페이지를 보여준다.")
+    void showLoginPage() {
+        webTestClient.get()
+                .uri("/login")
+                .exchange()
+                .expectStatus().isOk();
+    }
+}
